CSL: Finished [INTERLEAVE] rule
  * Had to change a transaction's semantics to fail if **any** of the participants
are not in the domain of the current state. Otherwise, if we allowed transaction
to create new entries with initial value 0, the following lemma does not hold anymore:
```agda
P `⊢ P′ → addr A P′ → addr A P
-- e.g.  P := A ↦ 1, P′ := (A ↦ 0 ∗ B ↦ 1) `∘⟦ A —→⟨ 1 ⟩ B ⟧
```
and you cannot fully prove [INTERLEAVE], for the case of `consequence`.

  * Also had to restrict from general monotone state transformers, to
just denotations of transactions (to make sure no new entries arise,
while monotonicity only disallowed **deleting** entries).

  * Still some minor postulates for equational reasoning conveniences
    - shorthands for transfering in doubleton maps
    - associativity/commutativity of _∗_